+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.2.1] - 2026-07-05
+
 ### 🐛 Fixed
 
 - `pulsar-scene-source`: repeated `pulsar-scene:SetCaptureSource` calls no